Fourth homework: implement book search by parameters using Specification API

+++ src/main/java/p/projects/springbookstore/service/impl/BookServiceImpl.java
@@ -2,15 +2,18 @@ package p.projects.springbookstore.service.impl;
 
 import java.util.List;
 import lombok.RequiredArgsConstructor;
+import org.springframework.data.jpa.domain.Specification;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import p.projects.springbookstore.dto.BookDto;
+import p.projects.springbookstore.dto.BookSearchParametersDto;
 import p.projects.springbookstore.dto.CreateBookRequestDto;
 import p.projects.springbookstore.dto.UpdateBookRequestDto;
 import p.projects.springbookstore.exception.EntityNotFoundException;
 import p.projects.springbookstore.mapper.BookMapper;
 import p.projects.springbookstore.model.Book;
 import p.projects.springbookstore.repository.BookRepository;
+import p.projects.springbookstore.repository.builder.BookSpecificationBuilder;
 import p.projects.springbookstore.service.BookService;
 
 @Service
@@ -18,6 +21,7 @@ import p.projects.springbookstore.service.BookService;
 public class BookServiceImpl implements BookService {
     private final BookRepository bookRepository;
     private final BookMapper bookMapper;
+    private final BookSpecificationBuilder specificationBuilder;
 
     @Override
     @Transactional
@@ -56,6 +60,18 @@ public class BookServiceImpl implements BookService {
     @Override
     @Transactional
     public void deleteBookById(Long id) {
+        if (!bookRepository.existsById(id)) {
+            throw new EntityNotFoundException("Can't find book with id: " + id);
+        }
         bookRepository.deleteById(id);
+    }
+
+    @Override
+    @Transactional(readOnly = true)
+    public List<BookDto> search(BookSearchParametersDto params) {
+        Specification<Book> bookSpecification = specificationBuilder.build(params);
+        return bookRepository.findAll(bookSpecification).stream()
+                .map(bookMapper::toDto)
+                .toList();
     }
 }
